+++ app/packages/utilities/src/index.ts
@@ -497,3 +497,36 @@ export const getMimeType = (sample: Sample) => {
   // mime type is null for certain file types like point-clouds
   return mimeFromFilePath ?? null;
 };
+
+export const toSlug = (name: string) => {
+  /**  Returns the URL-friendly slug for the given string.
+   *
+   * The following strategy is used to generate slugs:
+   *   (based on fiftyone.core.utils `to_slug` function)
+   *   -   The characters ``A-Za-z0-9`` are converted to lowercase
+   *   -   Whitespace and ``+_.-`` are converted to ``-``
+   *   -   All other characters are omitted
+   *   -   All consecutive ``-`` characters are reduced to a single ``-``
+   *   -   All leading and trailing ``-`` are stripped
+   */
+  if (name.length < 1) {
+    return "";
+  }
+  const valid_chars = new RegExp("[a-z0-9._+ -]", "g");
+  const replace_symbols = new RegExp("[-._+ ]+", "g");
+  const trim = new RegExp("-?(?<slug>[0-9a-z][0-9a-z-]*?)-?$");
+
+  let slug = name.toLowerCase();
+  let matches = [];
+  let match;
+  while ((match = valid_chars.exec(slug)) !== null) {
+    matches.push(match);
+  }
+  if (matches.length) {
+    slug = matches.join("")?.replace(replace_symbols, "-");
+    if (slug.length && slug !== "-") {
+      return slug.length ? trim.exec(slug)?.groups?.slug : "";
+    }
+  }
+  return "";
+};
